+++ tests/test_admin_ep.py
@@ -1,13 +1,59 @@
+import typing as t
+from enum import Enum
 import pytest
 
 pytest.importorskip("enum_properties")
 
-from tests.test_admin import TestAdmin
-from tests.enum_prop.models import AdminDisplayBug35
+from tests.test_admin import TestAdmin, TestEnumTesterAdminForm, _GenericAdminFormTest
+from tests.enum_prop.models import AdminDisplayBug35, EnumTester, BitFieldModel
+from tests.enum_prop.enums import (
+    GNSSConstellation,
+    LargeBitField,
+    LargeNegativeField,
+    LargeBitField,
+)
 
 
 class TestEnumPropAdmin(TestAdmin):
     BUG35_CLASS = AdminDisplayBug35
 
 
+class TestEnumTesterPropsAdminForm(TestEnumTesterAdminForm):
+    MODEL_CLASS = EnumTester
+    __test__ = True
+
+
+class TestBitFieldAdminForm(_GenericAdminFormTest):
+    MODEL_CLASS = BitFieldModel
+    HEADLESS = True
+    __test__ = True
+
+    @property
+    def changes(self) -> t.Dict[str, Enum]:
+        return [
+            {
+                "bit_field_small": GNSSConstellation.GLONASS | GNSSConstellation.GPS,
+                # "bit_field_large": LargeBitField.TWO | LargeBitField.ONE,
+                # "bit_field_large_empty_default": LargeBitField.TWO,
+                "large_neg": LargeNegativeField.NEG_ONE,
+                "no_default": LargeBitField.ONE | LargeBitField.TWO,
+            },
+            {
+                "bit_field_small": GNSSConstellation.GLONASS,
+                "bit_field_large": LargeBitField.TWO | LargeBitField.ONE,
+                "bit_field_large_empty_default": LargeBitField.TWO,
+                "large_neg": LargeNegativeField.ZERO,
+                "no_default": LargeBitField.TWO,
+            },
+            {
+                "bit_field_small": GNSSConstellation(0),
+                "bit_field_large": None,
+                "bit_field_large_empty_default": LargeBitField(0),
+                "large_neg": None,
+                "no_default": LargeBitField(0),
+            },
+        ]
+
+
 TestAdmin = None
+TestEnumTesterAdminForm = None